requirements.txt
asgiref==3.2.10
bootstrap4==0.1.0
Django==3.0.9
django-crispy-forms==1.9.2
django-registration-redux==2.8
mysqlclient==2.0.1
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
